real_4runr_deployment/4runr-outreach-system/requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyairtable==2.3.3
validators==0.22.0

# Web scraping and browser automation
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0

# AI/LLM integration
openai==1.30.5
httpx==0.27.0

# Template engine
jinja2==3.1.2

# Web service framework
fastapi==0.104.1
uvicorn==0.24.0

# Email validation and processing
email-validator==2.1.0

# Utilities and system tools
colorama==0.4.6
tqdm==4.66.1

# Development and testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1